Holy my goodness
So it turns out I attempted to refactor all of this a couple months ago,
got halfway through, then apparently gave up. So in preparation for a
trip on Sunday, I'm in full refactor-mode again. Changes in here:

- Pretty much writing everything from scratch
- "Support" for multiple providers (Spotify and Google Play Music, for
now)
- New non-Bootstrap (Skeleton) design and layout
- Moving to Font Awesome from Glyphicons
- Getting rid of the ridiculously overengineered npm and app/ stuff
- Getting rid of garbage custom Context objects
- Getting rid of garbage custom templates
- Adding tests for stuff!
- Supporting multiple rotation schemes (round robin vs random)

A bunch of stuff is still broken, but you can search for songs and add
them to a queue, and all of the room generation and joining should
hopefully be rock solid at this point.