A bunch of unrelated changes
- Remove choice of music servers
- Revamp (currently non-functional) *sqldb.DB (db.DB implementation) to
use a linked list of tracks to handle long track lists way better, and
also make insertions and deletions possible/less expensive.
- Redo how we do rotations, change the interface and update the tests
- Add the rng package, which is a copy of the Go pseudorandom source but
in a way that allows it to be serialized with Gob.
- Introduce a test harness and a basic test or two for the SQL
implementation. This should hopefully be generalizable to the memdb
implementation too, but that doesn't work right now.