Add the foundation for continuation tokens
One of the problems with the current Android app is that it has a
tendency to skip songs that were only partially played, because it sucks
at maintaining state. This change means that, for every song 'popped',
the server will return an opaque token that the client can (in the
future) use to request that song again.